Break out settings view, add netlify handlers
