+++ src/js/editor-view.tsx
@@ -1,14 +1,12 @@
 import styled from "@emotion/styled";
 import React, { FC, useCallback, useMemo, useState } from "react";
-import { ArrowLeft, Settings } from "react-feather";
-import { EditorContainer } from "./lib/styled";
+import { Settings } from "react-feather";
+import { Button, EditorContainer } from "./lib/styled";
 import { ObjectView } from "./object-view";
 import { ObjectTypesView } from "./object-types-view";
 import {
   Change,
   Github,
-  ObjectData,
-  ObjectDefinition,
   Objects,
   ObjectTypes,
   ProgressInfo,
@@ -19,6 +17,7 @@ import { ObjectsView } from "./objects-view";
 import { changeId } from "./lib/util";
 import { ErrorBoundary } from "react-error-boundary";
 import { ErrorView } from "./error-view";
+import { SettingsView } from "./settings-view";
 
 interface EditorViewProps {
   cloned: boolean;
@@ -50,79 +49,6 @@ interface EditorViewProps {
   cloneRepo: (branch: string) => void;
   reset: () => void;
 }
-
-const SettingsViewContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  font-size: 1.2em;
-  & > * {
-    margin-bottom: 1em;
-  }
-`;
-const SettingsHeader = styled.div`
-  display: flex;
-  align-items: center;
-  h2 {
-    flex-grow: 1;
-    font-size: 1em;
-  }
-`;
-
-interface SettingsViewProps {
-  cloning: boolean;
-  branch: string;
-  onReset: () => void;
-  onDismiss: () => void;
-  cloneRepo: (branch: string) => void;
-  progress: ProgressInfo | null;
-}
-
-const Button = styled.button`
-  border: 0;
-  border-radius: 5px;
-`;
-
-const ResetButton = styled(Button)`
-  background-color: red;
-  color: white;
-`;
-
-const SettingsView: FC<SettingsViewProps> = ({
-  branch: initialBranch,
-  cloning,
-  cloneRepo,
-  progress,
-  onReset,
-  onDismiss,
-}) => {
-  const [branch, updateBranch] = useState(initialBranch);
-  return (
-    <SettingsViewContainer>
-      <SettingsHeader>
-        <ArrowLeft onClick={onDismiss} />
-        <h2>Settings</h2>
-      </SettingsHeader>
-      <div>
-        {cloning ? null : (
-          <input
-            value={branch}
-            onChange={(e) => updateBranch(e.target.value)}
-            placeholder="branch"
-          />
-        )}
-        {cloning ? null : (
-          <Button onClick={() => cloneRepo(branch)}>Checkout Branch</Button>
-        )}
-      </div>
-      {progress ? (
-        <span>
-          {progress.task}: {Math.round(progress.progress * 100)}%
-        </span>
-      ) : null}
-      <ResetButton onClick={onReset}>Reset Repository</ResetButton>
-    </SettingsViewContainer>
-  );
-};
 
 const HeaderContainer = styled.div`
   display: flex;
